ref: Factor out some code into js-source-scopes (#682)
