--- symbolic-sourcemapcache/src/sourcemapcache/mod.rs
@@ -1,8 +1,0 @@
-#![allow(clippy::derive_partial_eq_without_eq)]
-
-mod lookup;
-mod raw;
-mod writer;
-
-pub use lookup::*;
-pub use writer::*;
